+++ sshd-common/src/test/java/org/apache/sshd/common/config/keys/writer/openssh/OpenSSHKeyPairResourceWriterTest.java
@@ -19,14 +19,6 @@
 
 package org.apache.sshd.common.config.keys.writer.openssh;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThrows;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
-
-import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -34,6 +26,7 @@ import java.io.StreamCorruptedException;
 import java.nio.ByteBuffer;
 import java.nio.channels.ByteChannel;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.security.GeneralSecurityException;
 import java.security.KeyPair;
@@ -43,9 +36,12 @@ import java.security.spec.AlgorithmParameterSpec;
 import java.security.spec.ECGenParameterSpec;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
 
+import net.i2p.crypto.eddsa.EdDSAKey;
+import net.i2p.crypto.eddsa.spec.EdDSAGenParameterSpec;
 import org.apache.sshd.common.config.keys.AuthorizedKeyEntry;
 import org.apache.sshd.common.config.keys.FilePasswordProvider;
 import org.apache.sshd.common.config.keys.KeyUtils;
@@ -53,86 +49,71 @@ import org.apache.sshd.common.config.keys.PublicKeyEntryResolver;
 import org.apache.sshd.common.util.io.SecureByteArrayOutputStream;
 import org.apache.sshd.common.util.io.resource.PathResource;
 import org.apache.sshd.common.util.security.SecurityUtils;
+import org.apache.sshd.util.test.JUnit4ClassRunnerWithParametersFactory;
+import org.apache.sshd.util.test.JUnitTestSupport;
+import org.apache.sshd.util.test.NoIoTestCase;
 import org.junit.Before;
+import org.junit.FixMethodOrder;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.experimental.categories.Category;
 import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
+import org.junit.runners.MethodSorters;
 import org.junit.runners.Parameterized;
-import org.junit.runners.Parameterized.Parameter;
 import org.junit.runners.Parameterized.Parameters;
+import org.junit.runners.Parameterized.UseParametersRunnerFactory;
 
-import net.i2p.crypto.eddsa.EdDSAKey;
-import net.i2p.crypto.eddsa.spec.EdDSAGenParameterSpec;
-
-@RunWith(Parameterized.class)
-public class OpenSSHKeyPairResourceWriterTest {
-
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+@RunWith(Parameterized.class) // see https://github.com/junit-team/junit/wiki/Parameterized-tests
+@UseParametersRunnerFactory(JUnit4ClassRunnerWithParametersFactory.class)
+@Category({ NoIoTestCase.class })
+public class OpenSSHKeyPairResourceWriterTest extends JUnitTestSupport {
     @Rule
     public TemporaryFolder folder = new TemporaryFolder();
 
-    private static class TestData {
-        public final String algorithm;
+    private final TestData data;
+    private KeyPair testKey;
 
-        public final String provider;
-
-        public final int keySize;
-
-        public final AlgorithmParameterSpec spec;
-
-        public TestData(String algorithm, int keySize,
-                AlgorithmParameterSpec spec) {
-            this(algorithm, null, keySize, spec);
-        }
-
-        public TestData(String algorithm, String provider, int keySize,
-                AlgorithmParameterSpec spec) {
-            this.algorithm = algorithm;
-            this.provider = provider;
-            this.keySize = keySize;
-            this.spec = spec;
-        }
-
-        @Override
-        public String toString() {
-            return algorithm + '-' + keySize
-                    + (provider == null ? "" : '(' + provider + ')');
-        }
+    public OpenSSHKeyPairResourceWriterTest(TestData data) {
+        this.data = data;
     }
 
     @Parameters(name = "{0}")
-    public static Object[] getParams() {
+    public static Collection<Object[]> parameters() {
         List<TestData> result = new ArrayList<>();
         result.add(new TestData("RSA", 1024, null));
         result.add(new TestData("RSA", 2048, null));
         result.add(new TestData("DSA", 1024, null));
-        result.add(
-                new TestData("ECDSA", 256,
-                        new ECGenParameterSpec("secp256r1")));
-        result.add(
-                new TestData("ECDSA", 384,
-                        new ECGenParameterSpec("secp384r1")));
-        result.add(
-                new TestData("ECDSA", 521,
-                        new ECGenParameterSpec("secp521r1")));
-        result.add(new TestData("EDDSA", "EdDSA", 25519,
-                new EdDSAGenParameterSpec("Ed25519")));
-        // Note: BC also has an EDDSA provider, but that one returns
-        // "Ed25519" as algorithm from its keys, while the one in
-        // net.i2p.crypto.eddsa gives keys with "EDDSA" as algorithm.
-        // sshd handles only the latter.
-        return result.toArray();
+        if (SecurityUtils.isECCSupported()) {
+            result.add(
+                    new TestData(
+                            "ECDSA", 256,
+                            new ECGenParameterSpec("secp256r1")));
+            result.add(
+                    new TestData(
+                            "ECDSA", 384,
+                            new ECGenParameterSpec("secp384r1")));
+            result.add(
+                    new TestData(
+                            "ECDSA", 521,
+                            new ECGenParameterSpec("secp521r1")));
+        }
+        if (SecurityUtils.isEDDSACurveSupported()) {
+            // Note: BC also has an EDDSA provider, but that one returns
+            // "Ed25519" as algorithm from its keys, while the one in
+            // net.i2p.crypto.eddsa gives keys with "EDDSA" as algorithm.
+            // sshd handles only the latter.
+            result.add(new TestData(
+                    "EDDSA", "EdDSA", 25519,
+                    new EdDSAGenParameterSpec("Ed25519")));
+        }
+
+        return parameterize(result);
     }
 
-    @Parameter
-    public TestData data;
-
-    private KeyPair testKey;
-
-    private OpenSSHKeyPairResourceWriter writer = new OpenSSHKeyPairResourceWriter();
-
     @Before
-    public void generateKey() throws Exception {
+    public void setUp() throws Exception {
         KeyPairGenerator generator;
         if (data.provider == null) {
             generator = KeyPairGenerator.getInstance(data.algorithm);
@@ -161,9 +142,9 @@ public class OpenSSHKeyPairResourceWriterTest {
         return KeyUtils.compareKeyPairs(a, b);
     }
 
-    private static void writeToFile(File file, byte[] sensitiveData)
+    private static void writeToFile(Path file, byte[] sensitiveData)
             throws IOException {
-        try (ByteChannel out = Files.newByteChannel(file.toPath(),
+        try (ByteChannel out = Files.newByteChannel(file,
                 StandardOpenOption.CREATE, StandardOpenOption.WRITE)) {
             ByteBuffer buf = ByteBuffer.wrap(sensitiveData);
             while (buf.hasRemaining()) {
@@ -175,103 +156,99 @@ public class OpenSSHKeyPairResourceWriterTest {
     }
 
     @Test
-    public void privateNoEncryption() throws Exception {
-        File tmp = folder.newFile();
+    public void testWritePrivateKeyNoEncryption() throws Exception {
+        Path tmp = getTemporaryOutputFile();
         try (SecureByteArrayOutputStream out = new SecureByteArrayOutputStream()) {
-            writer.writePrivateKey(testKey, "a comment", null, out);
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePrivateKey(testKey, "a comment", null, out);
             writeToFile(tmp, out.toByteArray());
         }
-        try (InputStream in = Files.newInputStream(tmp.toPath())) {
+        try (InputStream in = Files.newInputStream(tmp)) {
             KeyPair key = SecurityUtils.loadKeyPairIdentities(null,
-                    new PathResource(tmp.toPath()), in, null).iterator().next();
-            assertNotNull(key);
+                    new PathResource(tmp), in, null).iterator().next();
+            assertNotNull("No key pair parsed", key);
+            assertKeyPairEquals("Mismatched recovered keys", testKey, key);
             assertTrue("Keys should be equal", compare(key, testKey));
         }
     }
 
     @Test
-    public void privateNoPassword() throws Exception {
-        File tmp = folder.newFile();
-        OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+    public void testWritePrivateKeyNoPassword() throws Exception {
+        Path tmp = getTemporaryOutputFile();
         try (SecureByteArrayOutputStream out = new SecureByteArrayOutputStream()) {
-            writer.writePrivateKey(testKey, "a comment", options, out);
+            OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePrivateKey(testKey, "a comment", options, out);
             writeToFile(tmp, out.toByteArray());
-        } finally {
-            options.clear();
         }
-        try (InputStream in = Files.newInputStream(tmp.toPath())) {
+        try (InputStream in = Files.newInputStream(tmp)) {
             KeyPair key = SecurityUtils.loadKeyPairIdentities(null,
-                    new PathResource(tmp.toPath()), in, null).iterator().next();
-            assertNotNull(key);
+                    new PathResource(tmp), in, null).iterator().next();
+            assertNotNull("No key pair parsed", key);
+            assertKeyPairEquals("Mismatched recovered keys", testKey, key);
             assertTrue("Keys should be equal", compare(key, testKey));
         }
     }
 
     @Test
-    public void privateEncryptedAesCbc128() throws Exception {
-        File tmp = folder.newFile();
-        OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+    public void testWritePrivateKeyEncryptedAesCbc128() throws Exception {
+        Path tmp = getTemporaryOutputFile();
         try (SecureByteArrayOutputStream out = new SecureByteArrayOutputStream()) {
-            options.setPassphrase("nonsense".toCharArray());
+            OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+            options.setPassword("nonsense");
             options.setCipherName("AES");
             options.setCipherMode("CBC");
             options.setCipherType("128");
-            writer.writePrivateKey(testKey, "a comment", options, out);
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePrivateKey(testKey, "a comment", options, out);
             writeToFile(tmp, out.toByteArray());
-        } finally {
-            options.clear();
         }
-        try (InputStream in = Files.newInputStream(tmp.toPath())) {
+        try (InputStream in = Files.newInputStream(tmp)) {
             KeyPair key = SecurityUtils
-                    .loadKeyPairIdentities(null, new PathResource(tmp.toPath()),
+                    .loadKeyPairIdentities(null, new PathResource(tmp),
                             in, FilePasswordProvider.of("nonsense"))
                     .iterator().next();
-            assertNotNull(key);
+            assertNotNull("No key pair parsed", key);
+            assertKeyPairEquals("Mismatched recovered keys", testKey, key);
             assertTrue("Keys should be equal", compare(key, testKey));
         }
     }
 
     @Test
-    public void privateEncryptedAesCtr256() throws Exception {
-        File tmp = folder.newFile();
-        OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+    public void testWritePrivateKeyEncryptedAesCtr256() throws Exception {
+        Path tmp = getTemporaryOutputFile();
         try (SecureByteArrayOutputStream out = new SecureByteArrayOutputStream()) {
-            options.setPassphrase("nonsense".toCharArray());
+            OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+            options.setPassword("nonsense");
             options.setCipherName("AES");
             options.setCipherMode("CTR");
             options.setCipherType("256");
-            writer.writePrivateKey(testKey, "a comment", options, out);
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePrivateKey(testKey, "a comment", options, out);
             writeToFile(tmp, out.toByteArray());
-        } finally {
-            options.clear();
         }
-        try (InputStream in = Files.newInputStream(tmp.toPath())) {
+        try (InputStream in = Files.newInputStream(tmp)) {
             KeyPair key = SecurityUtils
-                    .loadKeyPairIdentities(null, new PathResource(tmp.toPath()),
+                    .loadKeyPairIdentities(null, new PathResource(tmp),
                             in, FilePasswordProvider.of("nonsense"))
                     .iterator().next();
-            assertNotNull(key);
+            assertNotNull("No key pair parsed", key);
+            assertKeyPairEquals("Mismatched recovered keys", testKey, key);
             assertTrue("Keys should be equal", compare(key, testKey));
         }
     }
 
     @Test
-    public void privateEncryptedWrongPassword() throws Exception {
-        File tmp = folder.newFile();
-        OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+    public void testWritePrivateKeyEncryptedWrongPassword() throws Exception {
+        Path tmp = getTemporaryOutputFile();
         try (SecureByteArrayOutputStream out = new SecureByteArrayOutputStream()) {
-            options.setPassphrase("nonsense".toCharArray());
+            OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+            options.setPassword("nonsense");
             options.setCipherName("AES");
             options.setCipherMode("CTR");
             options.setCipherType("256");
-            writer.writePrivateKey(testKey, "a comment", options, out);
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePrivateKey(testKey, "a comment", options, out);
             writeToFile(tmp, out.toByteArray());
-        } finally {
-            options.clear();
         }
-        try (InputStream in = Files.newInputStream(tmp.toPath())) {
+        try (InputStream in = Files.newInputStream(tmp)) {
             SecurityUtils.loadKeyPairIdentities(null,
-                    new PathResource(tmp.toPath()), in,
+                    new PathResource(tmp), in,
                     FilePasswordProvider.of("wrong"));
             fail("Expected an exception");
         } catch (StreamCorruptedException | GeneralSecurityException e) {
@@ -280,29 +257,26 @@ public class OpenSSHKeyPairResourceWriterTest {
     }
 
     @Test
-    public void privateEncryptedNoPassword() throws Exception {
-        File tmp = folder.newFile();
-        OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+    public void testWritePrivateKeyEncryptedNoPassword() throws Exception {
+        Path tmp = getTemporaryOutputFile();
         try (SecureByteArrayOutputStream out = new SecureByteArrayOutputStream()) {
-            options.setPassphrase("nonsense".toCharArray());
+            OpenSSHKeyEncryptionContext options = new OpenSSHKeyEncryptionContext();
+            options.setPassword("nonsense");
             options.setCipherName("AES");
             options.setCipherMode("CTR");
             options.setCipherType("256");
-            writer.writePrivateKey(testKey, "a comment", options, out);
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePrivateKey(testKey, "a comment", options, out);
             writeToFile(tmp, out.toByteArray());
-        } finally {
-            options.clear();
         }
-        try (InputStream in = Files.newInputStream(tmp.toPath())) {
+        try (InputStream in = Files.newInputStream(tmp)) {
             assertThrows(GeneralSecurityException.class,
                     () -> SecurityUtils.loadKeyPairIdentities(null,
-                            new PathResource(tmp.toPath()), in, null));
+                            new PathResource(tmp), in, null));
         }
     }
 
-    private void checkPublicKey(File tmp, String comment) throws Exception {
-        List<AuthorizedKeyEntry> keysRead = AuthorizedKeyEntry
-                .readAuthorizedKeys(tmp.toPath());
+    private void checkPublicKey(Path tmp, String comment) throws Exception {
+        List<AuthorizedKeyEntry> keysRead = AuthorizedKeyEntry.readAuthorizedKeys(tmp);
         assertEquals("Unexpected list size", 1, keysRead.size());
         AuthorizedKeyEntry entry = keysRead.get(0);
         String readComment = entry.getComment();
@@ -318,41 +292,76 @@ public class OpenSSHKeyPairResourceWriterTest {
     }
 
     @Test
-    public void publicWithComment() throws Exception {
-        File tmp = folder.newFile();
-        try (OutputStream out = new FileOutputStream(tmp)) {
-            writer.writePublicKey(testKey, "a comment", out);
+    public void testWritePublicKeyWithComment() throws Exception {
+        Path tmp = getTemporaryOutputFile();
+        try (OutputStream out = Files.newOutputStream(tmp)) {
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePublicKey(testKey, "a comment", out);
         }
         checkPublicKey(tmp, "a comment");
     }
 
     @Test
-    public void publicWithMultilineComment() throws Exception {
-        File tmp = folder.newFile();
-        try (OutputStream out = new FileOutputStream(tmp)) {
-            writer.writePublicKey(testKey,
+    public void testWritePublicKeyWithMultilineComment() throws Exception {
+        Path tmp = getTemporaryOutputFile();
+        try (OutputStream out = Files.newOutputStream(tmp)) {
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePublicKey(testKey,
                     "a comment" + System.lineSeparator() + "second line", out);
         }
         assertEquals("Unexpected number of lines", 1,
-                Files.readAllLines(tmp.toPath()).size());
+                Files.readAllLines(tmp).size());
         checkPublicKey(tmp, "a comment");
     }
 
     @Test
-    public void publicNoComment() throws Exception {
-        File tmp = folder.newFile();
-        try (OutputStream out = new FileOutputStream(tmp)) {
-            writer.writePublicKey(testKey, null, out);
+    public void testWritePublicKeyNoComment() throws Exception {
+        Path tmp = getTemporaryOutputFile();
+        try (OutputStream out = Files.newOutputStream(tmp)) {
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePublicKey(testKey, null, out);
         }
         checkPublicKey(tmp, null);
     }
 
     @Test
-    public void publicEmptyComment() throws Exception {
-        File tmp = folder.newFile();
-        try (OutputStream out = new FileOutputStream(tmp)) {
-            writer.writePublicKey(testKey, "", out);
+    public void testWritePublicKeyEmptyComment() throws Exception {
+        Path tmp = getTemporaryOutputFile();
+        try (OutputStream out = Files.newOutputStream(tmp)) {
+            OpenSSHKeyPairResourceWriter.INSTANCE.writePublicKey(testKey, "", out);
         }
         checkPublicKey(tmp, null);
+    }
+
+    private Path getTemporaryOutputFile() throws IOException {
+        Path dir = assertHierarchyTargetFolderExists(getTempTargetFolder());
+        return dir.resolve(getCurrentTestName());
+    }
+
+    @SuppressWarnings("checkstyle:VisibilityModifier")
+    private static class TestData {
+        public final String algorithm;
+
+        public final String provider;
+
+        public final int keySize;
+
+        public final AlgorithmParameterSpec spec;
+
+        TestData(String algorithm, int keySize,
+                 AlgorithmParameterSpec spec) {
+            this(algorithm, null, keySize, spec);
+        }
+
+        TestData(String algorithm, String provider, int keySize,
+                 AlgorithmParameterSpec spec) {
+            this.algorithm = algorithm;
+            this.provider = provider;
+            this.keySize = keySize;
+            this.spec = spec;
+        }
+
+        @Override
+        public String toString() {
+            return algorithm + '-' + keySize
+                   + (provider == null ? "" : '(' + provider + ')');
+        }
     }
 }